backened/code_editor_backened/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
cffi==1.17.1
channels==4.2.2
constantly==23.10.4
contourpy==1.1.1
cryptography==45.0.2
cycler==0.12.1
daphne==4.1.2
django==4.2.21
django-cors-headers==4.4.0
djangorestframework==3.15.2
fonttools==4.57.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
importlib-resources==6.4.5
incremental==24.7.2
kiwisolver==1.4.7
matplotlib==3.7.5
mysqlclient==2.2.7
numpy==1.24.4
packaging==25.0
pillow==10.4.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyopenssl==25.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.13.2
zipp==3.20.2
zope.interface==7.2
